Live /gemini-live-api-examples/gemini-live-ephemeral-tokens-websocket/pyproject.toml
[project]
name = "plain-js-demo-app"
version = "0.1.0"
description = "Gemini Live API - Vanilla JS Example"
dependencies = [
    "aiohttp>=3.8.0",
    "google-genai",
    "python-dotenv",
]

[tool.uv]
managed = true
